+++ muonShieldOptimization/study_thinTarget.py
@@ -4,7 +4,7 @@ ROOT.gROOT.ProcessLine('#include "FairModule.h"')
 time.sleep(20)
 
 import shipunit as u
-from ShipGeoConfig import ConfigRegistry
+import geometry_config
 
 mcEngine     = "TGeant4"
 runnr        = 1
@@ -28,7 +28,7 @@ ecut      = 0.0
 # -------------------------------------------------------------------
 ROOT.gRandom.SetSeed(theSeed)  # this should be propagated via ROOT to Pythia8 and Geant4VMC
 shipRoot_conf.configure()      # load basic libraries, prepare atexit for python
-ship_geo = ConfigRegistry.loadpy("$FAIRSHIP/geometry/geometry_config.py", Yheight = 10, tankDesign = 5, muShieldDesign = 7, nuTauTargetDesign=1)
+ship_geo = geometry_config.create_config(Yheight=10, shieldName="warm_opt")
 
 # -----Timer--------------------------------------------------------
 timer = ROOT.TStopwatch()